+++ Tests/Runtime/MediaStreamTrackTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ea866020d7e18ee40ac9f805785ba659
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
